feat: move vote proposal tx to shared
